requirements.txt
langchain
openai
python-dotenv
langchain-openai
langchain-community
langchain-core
langchain-google-genai